requirements.txt
africastalking==1.2.5
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.21
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
celery==5.3.2
redis==4.6.0
nltk==3.8.1
python-dateutil==2.8.2
phonenumbers==8.13.19
pydantic==2.4.2
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0
psycopg2-binary==2.9.7
azure-ai-inference==1.0.0b4
PyMySQL==1.1.0
mysql-connector-python==8.2.0
